made pause/unpause feature (same key)
i copied the code from p5.js and put it here. the way i did it was by having a variable, if its negative do things if its not, do nothing, and pressing the p key multiplies it by negative one and flips the value. like this:

if (key that was pressed === p) {
  isPaused  * -1  //flips from negative to positive or vice versa

if (isPaused === -1) {
  do stuff
}

if (isPaused === 1) {
  put text on screen saying its paused
}